Day 10, part 1
Pfiew, a bit head hurting; took me a while to see the button presses as
bit masks for the lights...
Had some help from Jipiti re. the bitmask creation for the buttons (I
had a working version, but it was rather ugly and I knew there had to be
a better way).