Day 10, documentation improvement and `print_machine` fix
Starting to ponder about part 2 and noticed `print_machine` wasn't
working anymore due to the rather weird handling of the buttons (ie. the
bitmasks they represent).
I reintroduced the length of the buttons, while changing the field's
name to a better one (it's not buttons anymore we're storing, but rather
the bitmasks they represent).  Also I think we might need the original
button data for part 2, in which case `buttons` would be a suitable
variable name.